Moved Pages into their own directories and common components into one folder.
