+++ src/common/components/atomic/Image.tsx
