+++ src/common/layouts/AutomaticLayout.tsx
