+++ src/pages/home/Home.tsx
@@ -1,33 +1,33 @@
 import React, { useState, useEffect } from "react"
 import { isBrowser, isMobile, BrowserView, MobileView } from "react-device-detect"
 
-import AutomaticLayout from "../layouts/AutomaticLayout"
+import AutomaticLayout from "../../common/layouts/AutomaticLayout"
 
-import Box from "../components/atomic/Box"
-import VStack from "../components/atomic/VStack"
-import HStack from "../components/atomic/HStack"
-import HDivider from "../components/atomic/HDivider"
-import Panel from "../components/atomic/Panel"
-import Badge from "../components/atomic/Badge"
-import TitleText from "../components/atomic/TitleText"
-import BodyText from "../components/atomic/BodyText"
-import Image from "../components/atomic/Image"
+import Box from "../../common/components/atomic/Box"
+import VStack from "../../common/components/atomic/VStack"
+import HStack from "../../common/components/atomic/HStack"
+import HDivider from "../../common/components/atomic/HDivider"
+import Panel from "../../common/components/atomic/Panel"
+import Badge from "../../common/components/atomic/Badge"
+import TitleText from "../../common/components/atomic/TitleText"
+import BodyText from "../../common/components/atomic/BodyText"
+import Image from "../../common/components/atomic/Image"
 
-import XStack from "../components/molecular/XStack"
-import ProfileImage from "../components/molecular/ProfileImage"
-import SectionPanel from "../components/molecular/SectionPanel"
-import FooterPanel from "../components/molecular/FooterPanel"
-import SocialMediaPanelButton from "../components/molecular/SocialMediaPanelButton"
+import XStack from "../../common/components/molecular/XStack"
+import ProfileImage from "../../common/components/molecular/ProfileImage"
+import SectionPanel from "../../common/components/molecular/SectionPanel"
+import FooterPanel from "../../common/components/molecular/FooterPanel"
+import SocialMediaPanelButton from "../../common/components/molecular/SocialMediaPanelButton"
 
-import DataContext from "../classes/DataContext"
-import Language from "../classes/Language"
-import Platform from "../classes/Platform"
-import Framework from "../classes/Framework"
+import DataContext from "../../classes/DataContext"
+import Language from "../../classes/Language"
+import Platform from "../../classes/Platform"
+import Framework from "../../classes/Framework"
 
-import initDataContext from "../data/initDataContext"
-import getAllLanguages from "../data/getAllLanguages"
-import getAllPlatforms from "../data/getAllPlatforms"
-import getAllFrameworks from "../data/getAllFrameworks"
+import initDataContext from "../../data/initDataContext"
+import getAllLanguages from "../../data/getAllLanguages"
+import getAllPlatforms from "../../data/getAllPlatforms"
+import getAllFrameworks from "../../data/getAllFrameworks"
 
 export default function Home() {
